requirements.txt
wheel
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
psycopg2-binary
Pillow
python-decouple
django-cors-headers
Django
qrcode
gunicorn
python-dateutil
schedule
celery
redis
config
twilio
